+++ test/test_all.hpp
@@ -1,6 +1,8 @@
 #ifndef REALM_TEST_TEST_ALL_HPP
 #define REALM_TEST_TEST_ALL_HPP
 
-int test_all(int argc, char* argv[]);
+#include <realm/util/logger.hpp>
+
+int test_all(int argc, char* argv[], realm::util::Logger* = nullptr);
 
 #endif // REALM_TEST_TEST_ALL_HPP